+++ Meowijuana/Screens/Testing/testing-andrea.cpp
@@ -27,8 +27,8 @@ void Andrea_Initialize()
 	Inventory::init();
 
 	// Setting position for inventory
-	float screenWidth = AEGfxGetWindowWidth();
-	float screenHeight = AEGfxGetWindowHeight();
+	float screenWidth = static_cast<float>(AEGfxGetWindowWidth());
+	float screenHeight = static_cast<float>(AEGfxGetWindowHeight());
 	float invWidth = inven.getSlotSize() * 9 + user.getInventorySize() * 8;
 	float offset = 10.0f;
